+++ src/types/arguments.rs
@@ -46,9 +46,10 @@ pub enum DeviceType {
 impl DeviceType {
     pub fn from_rdev(rdev: mode_t) -> Self {
         use libc::*;
+        let rdev: u32 = rdev.into();
         // Extract major and minor device numbers (assuming the device number format).
-        let major = rdev >> 8; // Major is the upper part of the 32-bit value
-        let minor = rdev & 0xFF; // Minor is the lower 8 bits
+        let major: u32 = rdev >> 8; // Major is the upper part of the 32-bit value (16 bit on macos)
+        let minor: u32 = rdev & 0xFF; // Minor is the lower 8 bits
         match rdev {
             x if x & S_IFREG != 0 => DeviceType::RegularFile,
             x if x & S_IFDIR != 0 => DeviceType::Directory,
@@ -67,8 +68,14 @@ impl DeviceType {
         match self {
             DeviceType::RegularFile => S_IFREG,
             DeviceType::Directory => S_IFDIR,
-            DeviceType::CharacterDevice { major, minor } => (major << 8) | (minor & 0xFF) | S_IFCHR,
-            DeviceType::BlockDevice { major, minor } => (major << 8) | (minor & 0xFF) | S_IFBLK,
+            DeviceType::CharacterDevice { major, minor } => {
+                let device = ((major & 0xFF) << 8) | (minor & 0xFF);
+                (device as mode_t) | S_IFCHR
+            },
+            DeviceType::BlockDevice { major, minor } => {
+                let device = ((major & 0xFF) << 8) | (minor & 0xFF);
+                (device as mode_t) | S_IFBLK
+            },
             DeviceType::NamedPipe => S_IFIFO,
             DeviceType::Socket => S_IFSOCK,
             DeviceType::Symlink => S_IFLNK,
