+++ src/unix_fs/bsd_like_fs.rs
@@ -1,9 +1,6 @@
-use crate::PosixError;
+use std::path::Path;
 
-use std::{ffi::c_void, path::Path};
-
-use crate::PosixError;
-use libc::{self, c_char, c_int, c_uint, off_t, size_t, ssize_t};
+use libc::{self, c_char, c_int, c_uint, off_t};
 
 use super::{cstring_from_path, FileDescriptor, StatFs};
 
@@ -32,34 +29,6 @@ pub(super) unsafe fn flush(fd: c_int) -> c_int {
 
 pub(super) unsafe fn fallocate(fd: c_int, offset: off_t, len: off_t) -> c_int {
     libc::posix_fallocate(fd, offset, len)
-}
-
-/// Retrieves file system statistics for the specified path.
-///
-/// This function is equivalent to the FUSE `statfs` operation.
-pub fn statfs(path: &Path) -> Result<StatFs, PosixError> {
-    let c_path = cstring_from_path(path)?;
-    let mut stat: libc::statfs = unsafe { std::mem::zeroed() };
-
-    // Use statfs to get file system stats
-    let result = unsafe { libc::statfs(c_path.as_ptr(), &mut stat) };
-    if result != 0 {
-        return Err(PosixError::last_error(format!(
-            "{}: statfs failed",
-            path.display()
-        )));
-    }
-
-    Ok(StatFs {
-        total_blocks: stat.f_blocks as u64,
-        free_blocks: stat.f_bfree as u64,
-        available_blocks: stat.f_bavail as u64,
-        total_files: stat.f_files as u64,
-        free_files: stat.f_ffree as u64,
-        block_size: stat.f_bsize as u32,
-        max_filename_length: stat.f_namemax as u32,
-        fragment_size: stat.f_bsize as u32, // BSD doesn't have f_frsize, so we use f_bsize
-    })
 }
 
 /// Copies a range of data from one file to another.
